Split command module into config and validation
