+++ src/lib.rs
@@ -4,7 +4,12 @@
 pub mod cli;
 
 /// Command configuration
-pub mod command;
+pub mod command {
+    /// Command config
+    pub mod config;
+    /// Validation utilities
+    pub mod validation;
+}
 
 /// Measurement functions
 pub mod measurement;
