+++ src/main.rs
@@ -3,7 +3,7 @@ use std::fmt::Debug;
 
 use clap::Parser;
 use git_perfdiff::cli::Args;
-use git_perfdiff::command::Config;
+use git_perfdiff::command::config::Config;
 use git_perfdiff::measurement::record_runtime;
 
 /// Print an error to stdout and exit with a failure code.
